requirements.txt
# Core Python packages
pillow>=9.0.0
mysql-connector-python>=8.0.30

# For development
pytest==7.4.3
flake8==6.1.0
black==23.11.0